+++ ai-backend/docker-compose.yml
@@ -7,6 +7,8 @@ services:
       target: dev
     ports:
       - "8000:8000"
+    env_file:
+      - .env
     environment:
       - LLM_API_BASE=http://194.12.189.112:11434
     volumes:
