--- ai-backend/src/config.py
@@ -1,1 +1,0 @@
-VALID_MODELS = ["llama3.2", "llama3.1:8b", "gemma2:27b"]
