+++ ai-backend/src/database.py
@@ -1,9 +1,9 @@
+from os import getenv
+
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from .config import DATABASE_URL
-
-engine = create_engine(DATABASE_URL)
+engine = create_engine(getenv("DATABASE_URL"))
 SessionLocal = sessionmaker(autocommit=False, autoflush=True, bind=engine)
