+++ ai-backend/src/main.py
@@ -1,7 +1,9 @@
+from os import getenv
+
 from fastapi import FastAPI
 
 from .api import v1
 
-app = FastAPI()
+app = FastAPI(root_path=getenv("ROOT_PATH", ""))
 
 app.include_router(v1.router, prefix="/v1")
