--- ai-backend/src/models/base.py
@@ -1,5 +1,0 @@
-from sqlalchemy.orm import DeclarativeBase
-
-
-class Base(DeclarativeBase):
-    pass
